Create ToggleComponent that can be used for all on/off archived and published toggle elements
