+++ src/js/views/programs/sidebar/flow/flow-sidebar_views.js
@@ -22,6 +22,7 @@ import Optionlist from 'js/components/optionlist';
 
 import { FlowBehaviorComponent, OwnerComponent } from 'js/views/programs/shared/flows_views';
 import TagsManagerComponent from 'js/views/programs/shared/components/tags-manager_component';
+import ToggleComponent from 'js/views/programs/shared/components/toggle_component';
 
 import FlowSidebarTemplate from './flow-sidebar.hbs';
 import FlowNameTemplate from './flow-name.hbs';
@@ -96,34 +97,6 @@ const DetailsView = View.extend({
     this.ui.spacer.text(text || ' ');
 
     this.model.set('details', trim(text));
-  },
-});
-
-const ToggleView = View.extend({
-  tagName: 'button',
-  attributes() {
-    return {
-      'disabled': this.getOption('isDisabled'),
-    };
-  },
-  className() {
-    const classNames = ['program-flow-sidebar__toggle button-secondary'];
-
-    if (this.getOption('status')) classNames.push('is-on');
-
-    return classNames.join(' ');
-  },
-  template: hbs`
-    {{#if status}}{{fas "toggle-on"}}{{else}}{{far "toggle-off"}}{{/if}}
-    {{formatMessage (intlGet "programs.sidebar.flow.flowSidebarViews.toggleView.toggle") status=status}}
-  `,
-  templateContext() {
-    return {
-      status: this.getOption('status'),
-    };
-  },
-  triggers: {
-    'click': 'click',
   },
 });
 
@@ -241,7 +214,7 @@ const LayoutView = View.extend({
     const published = !!this.flow.get('published_at');
     const isDisabled = this.flow.isNew();
 
-    const toggleView = new ToggleView({
+    const toggleView = new ToggleComponent({
       status: published,
       isDisabled,
     });
@@ -257,7 +230,7 @@ const LayoutView = View.extend({
     const archived = !!this.flow.get('archived_at');
     const isDisabled = this.flow.isNew();
 
-    const toggleView = new ToggleView({
+    const toggleView = new ToggleComponent({
       status: archived,
       isDisabled,
     });
